emails marked as checked but nothing received

+++ app/controllers/join_table_orders_items_controller.rb
@@ -1,5 +1,6 @@
 class JoinTableOrdersItemsController < ApplicationController
     before_action :create
+
     def index
     end
 
@@ -7,16 +8,25 @@ class JoinTableOrdersItemsController < ApplicationController
     end
 
     def create
-        @ord = Order.create(user_id: current_user.id)
-        @item = current_user.cart.items
-        @item.each do |item|
-        JoinTableOrdersItem.create(item_id: item.id, order_id: @ord.id)
-        end 
-        redirect_to root_path
-        current_user.cart.destroy
+        if current_user.cart.items.empty?
+            redirect_to cart_path(current_user.cart.id)
+        else
+            @ord = Order.create(user_id: current_user.id)
+            @item = current_user.cart.items
+            @item.each do |item|
+                JoinTableOrdersItem.create(item_id: item.id, order_id: @ord.id)
+            end 
+            order_validation_send
+            redirect_to root_path
+            current_user.cart.destroy
+        end
 
     end
 
     def destroy
+    end
+
+    def order_validation_send
+        UserMailer.order_validation_email(@ord).deliver_now
     end
 end
